requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
click==8.3.1
Django==6.0.4
django-cors-headers==4.9o
djangorestframework==3.17.1
fastapi==0.135.1
filelock==3.20.0
fsspec==2025.12.0
geographiclib==2.1
geopy==2.4.1
google-auth==2.49.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
-e git+ssh://git@github.com/PedroMoraes57/Reuse-Ai.git@5ba6ddf5f27873c5f9197f1ff8dfbf9d5a8a7295#egg=reuse_ai&subdirectory=backend
rich==14.3.3
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sqlparse==0.5.5
starlette==0.52.1
stringzilla==4.6.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.25
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
triton==3.1.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
